requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.6.0
Werkzeug==3.1.4
gunicorn==21.2.0
python-dotenv==1.2.1
